translate/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
PyMySQL==1.1.0
google-cloud-translate==3.12.1
python-dotenv==1.0.0